refactor: 삭제 쿼리 조건 추가

+++ backend/bottari/src/main/java/com/bottari/teambottari/service/TeamBottariService.java
@@ -149,7 +149,9 @@ public class TeamBottariService {
                 toDelete.add(info.getId());
             }
         }
-        teamAssignedItemInfoRepository.deleteByInfoIds(toDelete);
+        if (!toDelete.isEmpty()) {
+            teamAssignedItemInfoRepository.deleteByInfoIds(toDelete);
+        }
     }
 
     private Member getMemberBySsaid(final String ssaid) {
